+++ bin/amphora/src/amphora/node/transfer/mod.rs
@@ -1,0 +1,8 @@
+mod manager;
+mod pending;
+mod worker;
+
+use pending::{PendingTransfers, TransferGuard};
+use worker::TransferWorker;
+
+pub use manager::TransferManager;
